+++ R/utils-glmnet.R
@@ -1,3 +1,36 @@
+#' Avoid overlap in plotting ylabs
+#'
+#' Function to move ylabs up and down to avoid overlap of labels in lambda path
+#' for plot.rcv.glmnet(..., method = "path").
+#'
+#' @param y `numeric()`, y values.
+#' @param h `numeric(1)`, height of a single entry, `strheight` would be
+#' useful.
+#' @return `numeric()`, modified y.
+#' @noRd
+.avoid_ylab_overlap <- function(y, h) {
+    o <- order(y)
+    ys <- y[o]
+    n <- length(ys)
+    n1 <- n + 1L
+
+    step <- (h / 4)
+
+    d <- ys[-1L] - ys[-n]
+    iter <- 0L
+
+    while (any(d < h) && iter < 10L) {
+        w <- c(FALSE, d < h, FALSE)
+        ys[w[-1L]] <- ys[w[-1L]] - step
+        ys[w[-n1]] <- ys[w[-n1]] + step
+        d <- ys[-1L] - ys[-n]
+        iter <- iter + 1L
+    }
+
+    y[o] <- ys
+    y
+}
+
 #' Create balanced CV folds
 #'
 #' @param y `factor`, classes
@@ -41,6 +74,15 @@
 .mfolds <- function(y, nfolds = 3L, nrep = 5L, balanced = FALSE) {
     f <- if (isTRUE(balanced)) .bfolds else .folds
     do.call(rbind, lapply(integer(nrep), function(i)f(y, nfolds = nfolds)))
+}
+
+#' Nonzero Coefficients
+#'
+#' @param x `matrix`.
+#' @return `integer` row indices with nonzero beta coefficients from rcv.glmnet.
+#' @noRd
+.nonzero <- function(x) {
+    which(as.vector(x %*% rep(1L, ncol(x))) != 0)
 }
 
 #' Convert s into its numeric equivalent