+++ NEWS.md
@@ -2,6 +2,11 @@
 
 ## Changes in development
 
+## Changes in 0.0.25
+
+- Add `plot.rcvglmnet(..., what = "path")` to plot lambda path
+  (extends `glmnet::plot.glmnet`).
+
 ## Changes in 0.0.24
 
 - Pass `main` argument to `.plot.cal`, affected function: `plot.boot.glmnet`.